/ADD Ajout de l'initialisation du model ainsi que leur mise a jour en temps reel sur les boutons (Attention le cercle ne se complete si on relance l'app)

+++ StopCig/Model/Smoker_Model.swift
@@ -17,15 +17,15 @@ struct CigaretInfos : Codable {
 struct CigaretCountThisDay : Codable {
     var cigaretSmoked : Int
     var cigaretSaved : Int
-    var gain : Int
-    var lost : Int
+    var gain : Double
+    var lost : Double
 }
 
 struct CigaretTotalCount : Codable {
     var cigaretSmoked : Int
     var cigaretSaved : Int
-    var gain : Int
-    var lost : Int
+    var gain : Double
+    var lost : Double
 }
